+++ Userland/shell/commands/primes.c
@@ -8,7 +8,7 @@ typedef struct primesInternalState {
 
 static primeInternalState primeInternalStates[] = {{1}, {1}};
 
-static bool isPrime(int n) {
+static bool isPrime(int64_t n) {
     if(n <= 1) return FALSE;
     if(n <= 3) return TRUE;
     if(n % 2 == 0 || n % 3 == 0) return FALSE;
@@ -20,23 +20,23 @@ static bool isPrime(int n) {
     return TRUE;
 }
 
-static int calculateNextPrime(int n) {
+static int64_t calculateNextPrime(int64_t n) {
     do {
         n++;
     }while (!isPrime(n));
     return n;    
 }
 
-void primesNext(int screen) {
-    int n = calculateNextPrime(primeInternalStates[screen].number);
+void primesNext(int64_t screen) {
+    int64_t n = calculateNextPrime(primeInternalStates[screen].number);
     printf("-- %d\n", n);
     primeInternalStates[screen].number = n;
 }
 
-bool primesHasNext(int screen) {
+bool primesHasNext(int64_t screen) {
     return primeInternalStates[screen].number <=UINT64_MAX;
 }
 
-void primesReset(int screen) {
+void primesReset(int64_t screen) {
     primeInternalStates[screen].number = 1;
 }
