+++ Userland/shell/include/primes.h
@@ -3,8 +3,8 @@
 #include <stdint.h>
 #include <_stdbool.h>
 
-void primesNext(int screen);
-bool primesHasNext(int screen);
-void primesReset(int screen);
+void primesNext(int64_t screen);
+bool primesHasNext(int64_t screen);
+void primesReset(int64_t screen);
 
 #endif//_PRIMES_H
